Improve detecting script for iphone
Use xcode-select to do the work grabbing Xcode path for us. If anything
goes wrong, just goes back to default path as before.

This will accommodate for users that have multiple version of Xcode
installed on machine which is likely to happen for users that develop
for iOS platform, or iOS devs. It will select the active one set by such
tool xcode-select.